Move cart components from /features/cart to /features/ticketing/cart
